back/requirements.txt
asgiref==3.2.7
azure-cognitiveservices-vision-computervision==0.6.0
azure-cognitiveservices-vision-contentmoderator==1.0.0
azure-cognitiveservices-vision-face==0.4.0
azure-common==1.1.25
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==3.0.7
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.9
image==1.5.32
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.15
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
Pillow==7.1.2
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
